+++ src/index.ts
@@ -1,5 +1,5 @@
-export * from './Logger';
 export * from './lib/cyclic-serializer';
 export * from './loggers';
 export * as types from './types/';
 export * from './ComponentLogger';
+export * from './Logger';